formatter: weigh field names less when splitting groups by length
This adjusts how field lengths are calculated for the purposes of
determining whether to split groups upon encountering a long field.
For non-option fields, only 66% of the name length counts toward
the total length. This should make the formatter slightly more lenient
to long field names compared to long types.